requirements.txt
# NOMAD Threat Intelligence Framework Dependencies

# Core dependencies
feedparser>=6.0.10          # RSS/Atom feed parsing
pyyaml>=6.0                 # YAML configuration files
python-dateutil>=2.8.2      # Date/time handling
requests>=2.31.0            # HTTP requests for API calls

# LLM Integration (optional - for Claude Code integration)
anthropic>=0.18.0           # Anthropic API for Claude
openai>=1.12.0             # OpenAI API (alternative)

# Data validation
pydantic>=2.5.0            # Data validation and schemas
jsonschema>=4.20.0         # JSON schema validation

# Utilities
python-dotenv>=1.0.0       # Environment variable management
colorama>=0.4.6            # Colored terminal output
tabulate>=0.9.0            # Table formatting for CLI output

# Development dependencies
pytest>=7.4.0              # Testing framework
pytest-asyncio>=0.21.0     # Async test support
black>=23.12.0             # Code formatting
flake8>=6.1.0              # Linting
mypy>=1.7.0                # Type checking